requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot
google-genai>=0.3.0
yt-dlp>=2025.01.01
aiosqlite
pydantic
pydantic-settings
httpx>=0.27.0
ytmusicapi>=1.5.0
psutil
nest-asyncio
beautifulsoup4
spotipy
ffmpeg-python
brotli
pycryptodomex
pysocks
